requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.8
django-crispy-forms==1.13.0
django-phonenumber-field==5.2.0
phonenumbers==8.12.35
Pillow==8.4.0
pycodestyle==2.8.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
